pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-organizer"
version = "1.0.0"
description = "Offline face-recognition photo organizer — groups photos by person."
readme = "FACE_ORGANIZER_README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "face_recognition>=1.3.0",
    "dlib>=19.24",
    "Pillow>=10.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
face-organizer = "face_organizer.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
